+++ src/EFIGUI/Components/Input.cpp
@@ -13,13 +13,13 @@ namespace EFIGUI
 
     bool ModernInputText(const char* label, char* buf, size_t buf_size, ImGuiInputTextFlags flags)
     {
-        ImGui::PushStyleColor(ImGuiCol_FrameBg, Theme::ToVec4(Theme::ButtonDefault));
-        ImGui::PushStyleColor(ImGuiCol_FrameBgHovered, Theme::ToVec4(Theme::ButtonHover));
-        ImGui::PushStyleColor(ImGuiCol_FrameBgActive, Theme::ToVec4(Theme::ButtonHover));
-        ImGui::PushStyleColor(ImGuiCol_Border, Theme::ToVec4(Theme::BorderDefault));
+        ImGui::PushStyleColor(ImGuiCol_FrameBg, Theme::ToVec4(Theme::ButtonDefault()));
+        ImGui::PushStyleColor(ImGuiCol_FrameBgHovered, Theme::ToVec4(Theme::ButtonHover()));
+        ImGui::PushStyleColor(ImGuiCol_FrameBgActive, Theme::ToVec4(Theme::ButtonHover()));
+        ImGui::PushStyleColor(ImGuiCol_Border, Theme::ToVec4(Theme::BorderDefault()));
 
         ImGui::PushStyleVar(ImGuiStyleVar_FrameBorderSize, 1.0f);
-        ImGui::PushStyleVar(ImGuiStyleVar_FrameRounding, Theme::FrameRounding);
+        ImGui::PushStyleVar(ImGuiStyleVar_FrameRounding, Theme::FrameRounding());
 
         bool result = ImGui::InputText(label, buf, buf_size, flags);
 
@@ -31,9 +31,9 @@ namespace EFIGUI
 
     bool ModernInputTextMultiline(const char* label, char* buf, size_t buf_size, ImVec2 size, ImGuiInputTextFlags flags)
     {
-        ImGui::PushStyleColor(ImGuiCol_FrameBg, Theme::ToVec4(Theme::ButtonDefault));
-        ImGui::PushStyleColor(ImGuiCol_FrameBgHovered, Theme::ToVec4(Theme::ButtonHover));
-        ImGui::PushStyleColor(ImGuiCol_FrameBgActive, Theme::ToVec4(Theme::ButtonHover));
+        ImGui::PushStyleColor(ImGuiCol_FrameBg, Theme::ToVec4(Theme::ButtonDefault()));
+        ImGui::PushStyleColor(ImGuiCol_FrameBgHovered, Theme::ToVec4(Theme::ButtonHover()));
+        ImGui::PushStyleColor(ImGuiCol_FrameBgActive, Theme::ToVec4(Theme::ButtonHover()));
 
         bool result = ImGui::InputTextMultiline(label, buf, buf_size, size, flags);
 
@@ -60,13 +60,13 @@ namespace EFIGUI
         if (str->capacity() < 256)
             str->reserve(256);
 
-        ImGui::PushStyleColor(ImGuiCol_FrameBg, Theme::ToVec4(Theme::ButtonDefault));
-        ImGui::PushStyleColor(ImGuiCol_FrameBgHovered, Theme::ToVec4(Theme::ButtonHover));
-        ImGui::PushStyleColor(ImGuiCol_FrameBgActive, Theme::ToVec4(Theme::ButtonHover));
-        ImGui::PushStyleColor(ImGuiCol_Border, Theme::ToVec4(Theme::BorderDefault));
+        ImGui::PushStyleColor(ImGuiCol_FrameBg, Theme::ToVec4(Theme::ButtonDefault()));
+        ImGui::PushStyleColor(ImGuiCol_FrameBgHovered, Theme::ToVec4(Theme::ButtonHover()));
+        ImGui::PushStyleColor(ImGuiCol_FrameBgActive, Theme::ToVec4(Theme::ButtonHover()));
+        ImGui::PushStyleColor(ImGuiCol_Border, Theme::ToVec4(Theme::BorderDefault()));
 
         ImGui::PushStyleVar(ImGuiStyleVar_FrameBorderSize, 1.0f);
-        ImGui::PushStyleVar(ImGuiStyleVar_FrameRounding, Theme::FrameRounding);
+        ImGui::PushStyleVar(ImGuiStyleVar_FrameRounding, Theme::FrameRounding());
 
         flags |= ImGuiInputTextFlags_CallbackResize;
         bool result = ImGui::InputText(label, (char*)str->c_str(), str->capacity() + 1, flags, StringInputTextCallback, (void*)str);
@@ -82,9 +82,9 @@ namespace EFIGUI
         if (str->capacity() < 256)
             str->reserve(256);
 
-        ImGui::PushStyleColor(ImGuiCol_FrameBg, Theme::ToVec4(Theme::ButtonDefault));
-        ImGui::PushStyleColor(ImGuiCol_FrameBgHovered, Theme::ToVec4(Theme::ButtonHover));
-        ImGui::PushStyleColor(ImGuiCol_FrameBgActive, Theme::ToVec4(Theme::ButtonHover));
+        ImGui::PushStyleColor(ImGuiCol_FrameBg, Theme::ToVec4(Theme::ButtonDefault()));
+        ImGui::PushStyleColor(ImGuiCol_FrameBgHovered, Theme::ToVec4(Theme::ButtonHover()));
+        ImGui::PushStyleColor(ImGuiCol_FrameBgActive, Theme::ToVec4(Theme::ButtonHover()));
 
         flags |= ImGuiInputTextFlags_CallbackResize;
         bool result = ImGui::InputTextMultiline(label, (char*)str->c_str(), str->capacity() + 1, size, flags, StringInputTextCallback, (void*)str);